backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
openpyxl==3.1.5
python-multipart==0.0.12
reportlab==4.2.2
xlsxwriter==3.2.0
python-dateutil==2.9.0
msal==1.31.0
passlib[bcrypt]==1.7.4
